pyproject.toml
[project]
name = "projeto-fatura-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "streamlit>=1.40.0",
    "matplotlib>=3.7.0",
    "plotly>=5.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "prophet>=1.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]
